refactor: Move tests from 'unix' to 'daemon'
