+++ tests/daemon/clean.rs
