+++ tests/daemon/completions.rs
