+++ tests/daemon/edit.rs
