+++ tests/daemon/environment_variables.rs
