+++ tests/daemon/group.rs
