+++ tests/daemon/kill.rs
