+++ tests/daemon/log.rs
