+++ tests/daemon/mod.rs
