+++ tests/daemon/parallel_tasks.rs
