+++ tests/daemon/pause.rs
