+++ tests/daemon/remove.rs
