+++ tests/daemon/restart.rs
