+++ tests/daemon/restore.rs
