+++ tests/daemon/shutdown.rs
