+++ tests/daemon/start.rs
