+++ tests/daemon/stashed.rs
