+++ tests/daemon/worker_environment_variables.rs
